the script was throwing an error when there wasn't a uwsgi.conf so wrapped it with a check to make it more resiliant.

+++ 765ae2-stack/web/web.sh
@@ -73,6 +73,8 @@ sudo mkdir -p /media/data
 #
 # disable unused services
 #
-sudo mv /etc/supervisor/conf.d/uwsgi.conf /etc/supervisor/conf.d/uwsgi.save
+if [[ -f "/etc/supervisor/conf.d/uwsgi.conf" ]]; then
+    sudo mv /etc/supervisor/conf.d/uwsgi.conf /etc/supervisor/conf.d/uwsgi.save
+fi
 
 dcEndLog "install of app-specific web for 765ae2"
